Feature/TO-6487 area sub hubs activity trail

Approved-by: nouman

+++ app/Http/Controllers/Admins/ReturnController.php
@@ -4816,25 +4816,26 @@ class ReturnController extends Controller
         }
 
         $path = storage_path('app/public/uploads/return_notes');
-//        $paths = ['4', '5', '6', '7'];
-//        foreach ($paths as $p){
-            $files = File::glob("$path/2022*.*");
+        $paths = ['11', '12', '13', '14', '15', '16', '17', '18', '19', '20'];
+        foreach ($paths as $p){
+            $files = File::glob("$path/2023_$p*.*");
             $now = Carbon::now();
-
-            foreach ($files as $file) {
-                if (is_file($file)) {
-                    $created = date("F d Y H:i:s.",filemtime($file));
-                    $file_name = pathinfo($file);
-                    if($now->diffInDays($created) > 1){
-                        Storage::disk('s3')->put( 'return_note_images/'.$file_name['basename'], file_get_contents($file));
-                        $exists = Storage::disk('s3')->exists('return_note_images/'.$file_name['basename']);
-                        if($exists){
-                            File::delete($file);
+            if(count($files) > 0){
+                foreach ($files as $file) {
+                    if (is_file($file)) {
+                        $created = date("F d Y H:i:s.",filemtime($file));
+                        $file_name = pathinfo($file);
+                        if($now->diffInDays($created) > 1){
+                            Storage::disk('s3')->put( 'return_note_images/'.$file_name['basename'], file_get_contents($file));
+                            $exists = Storage::disk('s3')->exists('return_note_images/'.$file_name['basename']);
+                            if($exists){
+                                File::delete($file);
+                            }
                         }
                     }
                 }
             }
-//        }
+        }
     }
 
     public function cx_sales_index(){
